% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_classifier.R
\name{dummy_classifier}
\alias{dummy_classifier}
\title{dummy classifier to a categorical variable}
\usage{
dummy_classifier(
  strategy = "proportional",
  constant = NULL,
  random_state = NULL
)
}
\arguments{
\item{strategy}{a strategy from "constant", "most_frequent", "proportional", "uniform", or "stratified".}

\item{constant}{a constant value for the constant strategy.}

\item{random_state}{a random seed.}
}
\value{
a list
}
\description{
dummy classifier to a categorical variable
}
\examples{
model <- dummy_classifier(strategy = "uniform", random_state = 2024)(iris[,-5], iris$Species)
}
