% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_classifier_caret.R
\name{dummy_classifier_caret}
\alias{dummy_classifier_caret}
\title{dummy classifier for a categorical variable, used with the train function in caret.}
\usage{
dummy_classifier_caret(
  strategy = "proportional",
  constant = NULL,
  random_state = NULL
)
}
\arguments{
\item{strategy}{a strategy from "constant", "most_frequent", "proportional", "uniform", or "stratified".}

\item{constant}{a constant value for the constant strategy.}

\item{random_state}{a random seed.}
}
\value{
a list
}
\description{
dummy classifier for a categorical variable, used with the train function in caret.
}
