% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_regressor.R
\name{dummy_regressor}
\alias{dummy_regressor}
\title{dummy regressor}
\usage{
dummy_regressor(strategy = "mean", quantile = NULL, constant = NULL)
}
\arguments{
\item{strategy}{a strategy from "constant", "mean", "median", or "quantile".}

\item{quantile}{used when using the quantile strategy. It is a value between 0 and 1.}

\item{constant}{used when using the constant strategy. It is a numeric value.}
}
\value{
a list containing information of the model.
}
\description{
dummy regressor
}
\examples{
dummy_regressor(strategy = "mean")(iris[,-1], iris[,1])
}
