% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dummy_classifier.R
\name{predict_dummy_classifier}
\alias{predict_dummy_classifier}
\title{dummy classifier predictor}
\usage{
predict_dummy_classifier(object, X)
}
\arguments{
\item{object}{a list created using dummy classifier.}

\item{X}{a data frame.}
}
\value{
predicted values for the response variable.
}
\description{
dummy classifier predictor
}
\examples{
model <- dummy_classifier(strategy = "uniform", random_state = 2024)(iris[,-5], iris$Species)
predict_dummy_classifier(model, iris[,-5])
}
