% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dummy_regressor.R
\name{predict_dummy_regressor}
\alias{predict_dummy_regressor}
\title{dummy regressor predictor}
\usage{
predict_dummy_regressor(object, X)
}
\arguments{
\item{object}{a list from the dummy_regressor function}

\item{X}{a data frame}
}
\value{
the predicted values
}
\description{
dummy regressor predictor
}
\examples{
model <- dummy_regressor(strategy = "mean")(iris[,-1], iris[,1])
predict_dummy_regressor(model, iris[,-1])
}
