% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_proba.R
\name{predict_proba}
\alias{predict_proba}
\title{probabilities for predicting classes}
\usage{
predict_proba(model, X)
}
\arguments{
\item{model}{a list from dummy classifier.}

\item{X}{a data frame.}
}
\value{
a probability matrix.
}
\description{
probabilities for predicting classes
}
